requirements.txt
python-telegram-bot==20.0
rembg==2.0.62
Pillow==8.3.2
opencv-python==4.5.3.56